gf-system/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
captcha==0.6.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
discord.py==2.4.0
epicstore_api==0.1.9
frozenlist==1.4.1
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
multidict==6.0.5
pillow==10.4.0
pip==24.2
psutil==6.0.0
pyrsistent==0.20.0
python-dotenv==1.0.1
pytimeparse==1.1.8
rcon==2.4.6
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
soupsieve==2.6
urllib3==2.2.2
yarl==1.9.4
